adding model utils file for training dendritic networks
